Process no more than 1 mod per CPU at a time
Use Async.Parallel setting max parallelism to the number of CPU cores.

Simplify most of the hash calculation and mod lookup into an Async
workflow, eliminating most of the ModInfo messages that were previously
used.

Use a single message to communicate back to the main prosses the changes
in the state and progress for a mod.