Make Nexus a class
Avoid having to explicitly pass around and update the nexus record as
well as maintaining a separate ApiKey.

Turn the ApiKey into more of a domain object.  We know when it has been
validated and do not send an unvalidated API key to Nexus except when
explicitly requested, as in the validate call.

Move the Nir INI functions in to their own file instead of mixing
concerns in NexusApi.fs.

Add a bit of logic for displaying multiple results for an MD5 hash. This
as happens with some Septentrional Landscape files.