Fix pageHeader description position
Adding the buttons pushed the description line down.  This brings it
back up with the side-effect that if it is long, it will wrap before the
buttons instead of at the margin.  I may address this by direct
positioning later if necessary.

Refactor the code into different functions to make it easier to
rearrange things.

Move all the pageHeader styling to XAML.  In theory, this would make
restyling more flexible.