Handle Page/Msg mismatches
Extend the page history mechanism adding both forward and back
navigation. Create an "other" list to keep track of page state for pages
that are not in the linear history.

When a message arrives look up the related page in the history and other
lists.

As part of the history changes add a property to pages that indicate
whether or not they participate in history tracking or if they are modal
blocking all navigation via history or menu interfaces.  This is a
tri-state setting.