Restructure the build
Add an option (-i) to ignore code formatting issues.

Remove FSharpLint as part of the build.  There is still 'Lint' a fake
target.  It is called as part of the 'All' target and will be part of
the git pre-commit hook which will be added by a future commit.

Improve the fake dependencies so that a parallel build is possible and
runs more quickly.